requirements.txt
Flask==2.1.1
cryptography==36.0.1
qrcode==7.3.1
PyGithub==1.3.1
